requirements.txt
slack_bolt
python-dotenv
certifi
openai
langgraph
langgraph-sdk
langsmith
langchain-community
langchain-core
langchain-openai
langgraph-cli